requirements.txt
langchain
langchain-community
langchain-openai
langchain-astradb
astrapy
python-dotenv
django
pypdf
daphne
channels